requirements.txt
astropy==3.0.4
astropy-healpix==0.2.1
astroscrappy==1.0.5
ccdproc==1.3.0.post1
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.2
decorator==4.3.0
kiwisolver==1.0.1
matplotlib==3.0.0
networkx==2.2
numpy==1.15.2
photutils==0.5
Pillow==5.3.0
pyparsing==2.2.2
python-dateutil==2.7.3
PyWavelets==1.0.1
reproject==0.4
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
toolz==0.9.0
